Handle duplicated Vert.x context in CaffeineCacheImpl
